+++ scripts/lookup.sh
@@ -4,15 +4,16 @@ OUT_FILE="lookup.S"
 CSV_FILE="qrinfo.csv"
 SCRIPT_DIR="$(dirname "$0")"
 MAX_VERSION=40
+NUM_STEPS=5
 
 generate_info() {
-  echo "[1/4] Generating version info table"
+  printf "[1/%i] Generating version info table\n" "$NUM_STEPS"
   printf "\nglobale_ sym_(%s)\n" "qrinfo" >> "$OUT_FILE"
   python3 "$SCRIPT_DIR"/info.py "$SCRIPT_DIR"/"$CSV_FILE" >> "$OUT_FILE"
 }
 
 generate_indexes_data() {
-  echo "[2/4] Generating align pattern indexes data"
+  printf "[2/%i] Generating align pattern indexes data\n" "$NUM_STEPS"
   i=2
   while [ "$i" -le "$MAX_VERSION" ]; do
     printf "\nglobale_ sym_(%s_%i)\n" "qralign" "$i" >> "$OUT_FILE"
@@ -20,7 +21,15 @@ generate_indexes_data() {
     i=$((i + 1))
   done
 
-  echo "[3/4] Generating module indexes data"
+  printf "[3/%i] Generating masks xor pattern data\n" "$NUM_STEPS"
+  i=1
+  while [ "$i" -le "$MAX_VERSION" ]; do
+    printf "\nglobale_ sym_(%s_%i)\n" "qrmasks" "$i" >> "$OUT_FILE"
+    python3 "$SCRIPT_DIR"/indexes.py "$i" masks >> "$OUT_FILE"
+    i=$((i + 1))
+  done
+
+  printf "[4/%i] Generating module indexes data\n" "$NUM_STEPS"
   i=1
   while [ "$i" -le "$MAX_VERSION" ]; do
     printf "\nglobale_ sym_(%s_%i)\n" "qrindex" "$i" >> "$OUT_FILE"
@@ -30,7 +39,7 @@ generate_indexes_data() {
 }
 
 generate_pointers() {
-  echo "[4/4] Generating pointer tables"
+  printf "[5/%i] Generating pointer tables\n" "$NUM_STEPS"
   cat << 'EOF' >> "$OUT_FILE"
 
 #if defined(_WIN32) || defined(__MSYS__) || defined(__CYGWIN__)
@@ -46,6 +55,13 @@ EOF
   i=2
   while [ "$i" -le "$MAX_VERSION" ]; do
     printf "  pointer_ sym_(%s_%i)\n" "qralign" "$i" >> "$OUT_FILE"
+    i=$((i + 1))
+  done
+
+  printf "\nglobale_ sym_(%s)\n" "qrmasks" >> "$OUT_FILE"
+  i=1
+  while [ "$i" -le "$MAX_VERSION" ]; do
+    printf "  pointer_ sym_(%s_%i)\n" "qrmasks" "$i" >> "$OUT_FILE"
     i=$((i + 1))
   done