+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Add XOR masks pre-calculated LUT
+
 ## [1.20.0] - 2026-03-09
 
 Add full Version and Level support